@chg 	Changed version of junit 4.1-->4.4. 	As a consequence, 	- had to replace assertEquals(double, double) 	  by assertEquals(double, double, 0.0) very frequently. 	- for assertEquals(int, int) had to ensure that both arguments are really int not Integer 	 (otherwise may be both Integer)
git-svn-id: file:///SVN_Archive/Software/trunk@6291 9b6744e0-6d27-0410-8403-e1b9a971d29b

+++ src/test/java/eu/simuline/util/BitSetListTest.java
@@ -316,10 +316,10 @@ assertEquals(4,bitSetList.size());
 	bitSetList.add(0);
 	bitSetList.add(0);
 
-	assertEquals(1,bitSetList.get(0));
-	assertEquals(1,bitSetList.get(1));
-	assertEquals(0,bitSetList.get(2));
-	assertEquals(0,bitSetList.get(3));
+	assertEquals(1,bitSetList.get(0).intValue());
+	assertEquals(1,bitSetList.get(1).intValue());
+	assertEquals(0,bitSetList.get(2).intValue());
+	assertEquals(0,bitSetList.get(3).intValue());
 	try {
 	    bitSetList.get(-1);
 	    fail();
